Adding package creator script
